# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  queued: 'queued',
  processing: 'processing',
  done: 'done',
  failed: 'failed',
  cancelled: 'cancelled',
} as const;

export interface Job {
  id: string;
  prompt: string;
  duration: number;
  resolutionWidth: number;
  resolutionHeight: number;
  status: JobStatus;
  progress: number;
  retryCount: number;
  maxRetries: number;
  priority: number;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  lockedAt?: string | null;
  /** @nullable */
  lockedBy?: string | null;
  /** @nullable */
  currentStage?: string | null;
  /** @nullable */
  presetName?: string | null;
  referenceImages: string[];
  characterIds: string[];
}

export interface JobAttempt {
  id: string;
  jobId: string;
  attemptNumber: number;
  startedAt: string;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  durationMs?: number | null;
  /** @nullable */
  result?: string | null;
  /** @nullable */
  errorMessage?: string | null;
}

export interface SystemStatus {
  worker: string;
  workerId: string;
  concurrency: number;
  queue: number;
  processing: number;
  failed: number;
  uptime: string;
  /** @nullable */
  lastHeartbeat: string | null;
  currentJobIds: string[];
  memoryMB: number;
}

export interface SystemMetrics {
  totalJobs: number;
  success: number;
  failed: number;
  cancelled: number;
  /** @nullable */
  avgProcessingTimeMs: number | null;
  /** @nullable */
  avgProcessingTime: string | null;
  successRate: string;
}

export type AssetType = typeof AssetType[keyof typeof AssetType];


export const AssetType = {
  input_image: 'input_image',
  generated_video: 'generated_video',
  thumbnail: 'thumbnail',
  audio: 'audio',
  subtitle: 'subtitle',
} as const;

export interface Asset {
  id: string;
  jobId: string;
  type: AssetType;
  path: string;
  url: string;
  size: number;
  mime: string;
  createdAt: string;
}

export interface JobAssets {
  assets: Asset[];
  /** @nullable */
  video: string | null;
  /** @nullable */
  thumbnail: string | null;
}

export interface GenerationPreset {
  id: string;
  name: string;
  resolution: string;
  fps: number;
  style: string;
  /** @nullable */
  description?: string | null;
  isDefault: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePresetRequest {
  name: string;
  resolution?: string;
  fps?: number;
  style?: string;
  description?: string;
  isDefault?: boolean;
}

export interface Character {
  id: string;
  name: string;
  /** @nullable */
  referenceImage?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  embedding?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCharacterRequest {
  name: string;
  referenceImage?: string;
  description?: string;
}

export interface UpdateCharacterRequest {
  name?: string;
  referenceImage?: string;
  description?: string;
}

export interface UploadResponse {
  url: string;
  path: string;
  size: number;
  mime: string;
}

export interface CreateJobRequest {
  prompt: string;
  duration?: number;
  resolutionWidth?: number;
  resolutionHeight?: number;
  priority?: number;
  preset?: string;
  imageReferences?: string[];
  characterIds?: string[];
}

export type WebhookEvent = typeof WebhookEvent[keyof typeof WebhookEvent];


export const WebhookEvent = {
  jobdone: 'job.done',
  jobfailed: 'job.failed',
  joball: 'job.all',
} as const;

export interface Webhook {
  id: string;
  url: string;
  event: WebhookEvent;
  /** @nullable */
  label?: string | null;
  createdAt: string;
  /** @nullable */
  lastFiredAt?: string | null;
  /** @nullable */
  lastStatusCode?: number | null;
}

export interface WebhookInput {
  url: string;
  event: WebhookEvent;
  label?: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListJobs200 = {
  jobs: Job[];
};

export type ListJobAttempts200 = {
  attempts: JobAttempt[];
};

export type ListPresets200 = {
  presets: GenerationPreset[];
};

export type ListCharacters200 = {
  characters: Character[];
};

export type ListWebhooks200 = {
  webhooks: Webhook[];
};

export type TestWebhook200 = {
  ok: boolean;
  statusCode: number;
};