# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all jobs
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: "#/components/schemas/Job"
                required:
                  - jobs
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a video generation job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/{id}/attempts:
    get:
      operationId: listJobAttempts
      tags: [jobs]
      summary: List attempt history for a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of attempts
          content:
            application/json:
              schema:
                type: object
                properties:
                  attempts:
                    type: array
                    items:
                      $ref: "#/components/schemas/JobAttempt"
                required:
                  - attempts
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /characters:
    get:
      operationId: listCharacters
      tags: [characters]
      summary: List all characters
      responses:
        "200":
          description: List of characters
          content:
            application/json:
              schema:
                type: object
                properties:
                  characters:
                    type: array
                    items:
                      $ref: "#/components/schemas/Character"
                required:
                  - characters
    post:
      operationId: createCharacter
      tags: [characters]
      summary: Create a character
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCharacterRequest"
      responses:
        "201":
          description: Character created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
  /characters/{id}:
    get:
      operationId: getCharacter
      tags: [characters]
      summary: Get a character by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Character found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "404":
          description: Character not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCharacter
      tags: [characters]
      summary: Update a character
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCharacterRequest"
      responses:
        "200":
          description: Character updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "404":
          description: Character not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCharacter
      tags: [characters]
      summary: Delete a character
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Character deleted
        "404":
          description: Character not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /jobs/{id}/assets:
    get:
      operationId: listJobAssets
      tags: [jobs]
      summary: List assets produced by a completed job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Assets for the job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAssets"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/{id}/retry:
    post:
      operationId: retryJob
      tags: [jobs]
      summary: Retry a failed or cancelled job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job re-queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Cannot retry in current state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/{id}/cancel:
    post:
      operationId: cancelJob
      tags: [jobs]
      summary: Cancel a queued or processing job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job cancelled (or cancellation signal sent for processing jobs)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Cannot cancel in current state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    JobStatus:
      type: string
      enum:
        - queued
        - processing
        - done
        - failed
        - cancelled
    Job:
      type: object
      properties:
        id:
          type: string
        prompt:
          type: string
        duration:
          type: integer
        resolutionWidth:
          type: integer
        resolutionHeight:
          type: integer
        status:
          $ref: "#/components/schemas/JobStatus"
        progress:
          type: integer
        retryCount:
          type: integer
        maxRetries:
          type: integer
        priority:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        errorMessage:
          type: ["string", "null"]
        lockedAt:
          type: ["string", "null"]
        lockedBy:
          type: ["string", "null"]
        currentStage:
          type: ["string", "null"]
        presetName:
          type: ["string", "null"]
        referenceImages:
          type: array
          items:
            type: string
        characterIds:
          type: array
          items:
            type: string
      required:
        - id
        - prompt
        - duration
        - resolutionWidth
        - resolutionHeight
        - status
        - progress
        - retryCount
        - maxRetries
        - priority
        - referenceImages
        - characterIds
        - createdAt
        - updatedAt
    JobAttempt:
      type: object
      properties:
        id:
          type: string
        jobId:
          type: string
        attemptNumber:
          type: integer
        startedAt:
          type: string
        completedAt:
          type: ["string", "null"]
        durationMs:
          type: ["integer", "null"]
        result:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
      required:
        - id
        - jobId
        - attemptNumber
        - startedAt
...
    Asset:
      type: object
      properties:
        id:
          type: string
        jobId:
          type: string
        type:
          type: string
          enum: [input_image, generated_video, thumbnail, audio, subtitle]
        path:
          type: string
        url:
          type: string
        size:
          type: integer
        mime:
          type: string
        createdAt:
          type: string
      required:
        - id
        - jobId
        - type
        - path
        - url
        - size
        - mime
        - createdAt
    JobAssets:
      type: object
      properties:
        assets:
          type: array
          items:
            $ref: "#/components/schemas/Asset"
        video:
          type: ["string", "null"]
        thumbnail:
          type: ["string", "null"]
      required:
        - assets
        - video
        - thumbnail
...
    Character:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        referenceImage:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        embedding:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - createdAt
        - updatedAt
    CreateCharacterRequest:
      type: object
      properties:
        name:
          type: string
        referenceImage:
          type: string
        description:
          type: string
      required:
        - name
    UpdateCharacterRequest:
      type: object
      properties:
        name:
          type: string
        referenceImage:
          type: string
        description:
          type: string
...
    CreateJobRequest:
      type: object
      properties:
        prompt:
          type: string
        duration:
          type: integer
          default: 5
        resolutionWidth:
          type: integer
          default: 720
        resolutionHeight:
          type: integer
          default: 1280
        priority:
          type: integer
          default: 0
        preset:
          type: string
        imageReferences:
          type: array
          items:
            type: string
        characterIds:
          type: array
          items:
            type: string
      required:
        - prompt
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error