# lib/api-zod/src/generated/types/character.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Character {
  id: string;
  name: string;
  /** @nullable */
  referenceImage?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  embedding?: string | null;
  createdAt: string;
  updatedAt: string;
}