# lib/api-zod/src/generated/types/createJobRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateJobRequest {
  prompt: string;
  duration?: number;
  resolutionWidth?: number;
  resolutionHeight?: number;
  priority?: number;
  preset?: string;
  imageReferences?: string[];
  characterIds?: string[];
}