# lib/api-zod/src/generated/types/job.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobStatus } from './jobStatus';

export interface Job {
  id: string;
  prompt: string;
  duration: number;
  resolutionWidth: number;
  resolutionHeight: number;
  status: JobStatus;
  progress: number;
  retryCount: number;
  maxRetries: number;
  priority: number;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  lockedAt?: string | null;
  /** @nullable */
  lockedBy?: string | null;
  /** @nullable */
  currentStage?: string | null;
  /** @nullable */
  presetName?: string | null;
  referenceImages: string[];
  characterIds: string[];
}