# lib/api-zod/src/generated/types/updateCharacterRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateCharacterRequest {
  name?: string;
  referenceImage?: string;
  description?: string;
}